requirements.txt
quart
sqlalchemy
python-dotenv
aiomysql